backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
greenlet>=3.0.0
psycopg2-binary==2.9.10
alembic==1.14.0

# HTTP / scraping
httpx==0.28.0
feedparser==6.0.11

# AI
google-genai==1.0.0

# Config / utilities
python-dotenv==1.0.1
pydantic==2.10.0
pydantic-settings==2.7.0
